+++ AmdarisProject/AmdarisProject.Application/Dtos/ResponseDTOs/GetDTOs/CompetitionGetDTOs/CompetitionGetDTO.cs
@@ -13,7 +13,7 @@ namespace AmdarisProject.Application.Dtos.ResponseDTOs.CompetitionResponseDTOs
         public required ulong? BreakInMinutes { get; set; }
         public required GameType GameType { get; set; }
         public required CompetitorType CompetitorType { get; set; }
-        public required ushort? TeamSize { get; set; }
+        public required uint? TeamSize { get; set; }
         public required uint? WinAt { get; set; }
         public required ulong? DurationInMinutes { get; set; }
         public required List<CompetitorDisplayDTO> Competitors { get; set; } = [];
