+++ AmdarisProject/AmdarisProject.Application/Dtos/ResponseDTOs/GetDTOs/CompetitionGetDTOs/TournamentCompetitionGetDTO.cs
@@ -2,6 +2,6 @@
 {
     public class TournamentCompetitionGetDTO : CompetitionGetDTO
     {
-        public required ushort StageLevel { get; set; }
+        public required uint StageLevel { get; set; }
     }
 }
