+++ AmdarisProject/AmdarisProject.Domain/Models/CompetitionModels/TournamentCompetition.cs
@@ -2,6 +2,6 @@
 {
     public class TournamentCompetition : Competition
     {
-        public required ushort StageLevel { get; set; }
+        public required uint StageLevel { get; set; }
     }
 }
