+++ AmdarisProject/AmdarisProject.Domain/Models/GameFormat.cs
@@ -7,7 +7,7 @@ namespace AmdarisProject.Domain.Models
         public required string Name { get; set; }
         public required GameType GameType { get; set; }
         public required CompetitorType CompetitorType { get; set; }
-        public required ushort? TeamSize { get; set; }
+        public required uint? TeamSize { get; set; }
         public required uint? WinAt { get; set; }
         public required ulong? DurationInMinutes { get; set; }
     }
